Correcting some mistakes on tutorial 1
the line "from reservoirpy.nodes import Reservoir, Ridge, FORCE, ESN" doesn't seems important. Removing it doesn't change anything.
There was a mistake at "bu" for "but" in the phrase "Then, we train the readout using fit(). We use the warmup parameter to set the amount of timesteps we want to discard at the beginning of train_states, when training. Indeed, as the internal state of the reservoir is recurrently updated bu start to a null value, it takes time before this state is representative of the input data."

It seems that removing an entire block of code change a lot of lines. The difference seems huge for not a lot of reasons.