Correction of some minor mistakes on the tutorials (#161)
* Correcting some mistakes on tutorial 6

There was one mistake on the 'i' of 'instances'. 'ìnstances' => 'instances'

* Correcting some mistakes on tutorial 5

There was missing an 'e' on 'Echo state networks'. 'Echo stat networks' => 'Echo state networks'. Some comments on a code had too many line break. Finaly the was a gramatical error. 'With split' => 'We split'

* Correcting some mistakes on tutorial 3

There was a line break were none were required : 'plt.legend(
)'=> 'plt.legend()

The parameter sparse for the OneHotEncore function is now sparse_output : 'sparse=False'=>'sparse_output=False'

* Correcting some mistakes on tutorial 1

the line "from reservoirpy.nodes import Reservoir, Ridge, FORCE, ESN" doesn't seems important. Removing it doesn't change anything.
There was a mistake at "bu" for "but" in the phrase "Then, we train the readout using fit(). We use the warmup parameter to set the amount of timesteps we want to discard at the beginning of train_states, when training. Indeed, as the internal state of the reservoir is recurrently updated bu start to a null value, it takes time before this state is representative of the input data."

It seems that removing an entire block of code change a lot of lines. The difference seems huge for not a lot of reasons.

* Doc: non-breaking space + removed image diff

---------

Co-authored-by: PAUL BERNARD <Paul.Bernard@inria.fr>